prep.dirs: generalize ceph repo template creation
If ceph_repo_base_url is given, use that as the ceph repo base url.

If ceph_repo_base_url is not given, query shaman to find the latest ceph main
dev build and use it as the ceph repo base url.

A ceph_repo_gpgkey variable can also be given to allow for gpg signed repos.

Signed-off-by: Andrew Schoen <aschoen@ibm.com>