Add new third_party libraries for SMB over QUIC
These are created based on sources from the following repositories:
- https://github.com/lxin/quic
- https://github.com/ngtcp2/ngtcp2

Since the kernel implementation is still not merged there's also a
wrapper which is only built when configured with `--enable-selftest`.
For now it is ignored as the wrapper may finally end up in its own
repository than Samba in the near future.

ref: https://gitlab.com/samba-team/samba/-/merge_requests/4019