rpmspec: Conditionalize the installation of libcmocka-devel
cmocka in Samba was recently upgraded[1] to v1.1.8 but it is not
available with EL9 and EL10 making a fallback to the third_party
sources creating an additional samba specific shared library to be
included during package build process. For now conditionally install
libcmocka-devel on Fedora version > 41 and any EL version > 10 (because
cmocka v1.1.8 is already built for ELN which becomes the next RHEL).

[1] https://gitlab.com/samba-team/samba/-/merge_requests/4269